se modifica motores de busqueda de libros, por titulo de libros

+++ app/Http/Controllers/PostController.php
@@ -75,9 +75,13 @@ class PostController extends Controller
     {
 
         $comentarios = Comment::where('post_id',$id['id'])->get();
-        foreach($comentarios as $come){
-            $come->delete();
+        dd($comentarios);
+        if(!empty($comentarios)){
+            foreach($comentarios as $come){
+                $come->delete();
+            }
         }
+        
        
 
        $post=Post::find($id['id']);
